# internal/api/swagger/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Watchtower HTTP API for container update management, metrics, and health probes.",
        "title": "Watchtower HTTP API",
        "contact": {
            "name": "Nicholas Fedor",
            "url": "https://github.com/sidneyojr/watchtower"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/v1/containers/details": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns detailed information about each watched container, including running state, image identity, and configuration flags. Optionally filter by container name or image name.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "containers"
                ],
                "summary": "Detailed container status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by container name (exact match)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by image name (exact match)",
                        "name": "image",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Container details with count and timestamp",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API token",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to get container details",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Paste the HTTP API token only (WATCHTOWER_HTTP_API_TOKEN). Swagger UI sends it as the Authorization header value. curl clients should use \"Authorization: Bearer \u003ctoken\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "EventsToken": {
            "description": "Paste the events API token only (WATCHTOWER_HTTP_API_EVENTS_TOKEN). Swagger UI sends it as the Authorization header value.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}